탐색 탭 API 응답 언어 라벨 추가
